@@ -52,45 +55,48 @@
     "prop-types": "^15.6.0"
   },
   "devDependencies": {
-    "@trendmicro/react-buttons": "~1.3.0",
-    "babel-cli": "~6.26.0",
-    "babel-core": "~6.26.0",
-    "babel-eslint": "~8.2.2",
-    "babel-loader": "~7.1.4",
-    "babel-plugin-transform-decorators-legacy": "~1.3.4",
-    "babel-preset-env": "^1.6.1",
-    "babel-preset-react": "~6.24.1",
-    "babel-preset-stage-0": "~6.24.1",
+    "@babel/cli": "~7.2.3",
+    "@babel/core": "~7.2.2",
+    "@babel/polyfill": "~7.2.3",
+    "@babel/preset-env": "~7.2.3",
+    "@babel/preset-react": "~7.0.0",
+    "@babel/register": "~7.0.0",
+    "@trendmicro/babel-config": "~1.0.0-alpha",
+    "@trendmicro/react-buttons": "~1.3.1",
+    "babel-eslint": "~10.0.1",
+    "babel-loader": "~8.0.4",
     "chained-function": "~0.5.0",
-    "clean-css": "~4.1.9",
-    "clean-css-cli": "~4.1.10",
-    "coveralls": "~3.0.0",
-    "css-loader": "~0.28.9",
+    "clean-css": "~4.2.1",
+    "clean-css-cli": "~4.2.1",
+    "coveralls": "~3.0.2",
+    "css-loader": "~2.1.0",
     "enzyme": "~3.8.0",
     "enzyme-adapter-react-16": "~1.7.1",
-    "eslint": "~4.18.2",
-    "eslint-config-trendmicro": "~1.3.0",
-    "eslint-loader": "~2.0.0",
-    "eslint-plugin-import": "~2.9.0",
-    "eslint-plugin-jsx-a11y": "~6.0.3",
-    "eslint-plugin-react": "~7.7.0",
-    "extract-text-webpack-plugin": "~3.0.2",
-    "file-loader": "~1.1.6",
-    "find-imports": "~0.5.2",
-    "html-webpack-plugin": "~3.0.6",
+    "eslint": "~5.11.1",
+    "eslint-config-trendmicro": "~1.4.1",
+    "eslint-loader": "~2.1.1",
+    "eslint-plugin-import": "~2.14.0",
+    "eslint-plugin-jsx-a11y": "~6.1.2",
+    "eslint-plugin-react": "~7.12.0",
+    "file-loader": "~3.0.1",
+    "find-imports": "^1.1.0",
+    "html-webpack-plugin": "~3.2.0",
     "http-server": "~0.11.1",
-    "jsdom": "~11.6.0",
+    "jsdom": "^13.1.0",
+    "mini-css-extract-plugin": "~0.5.0",
     "nib": "~1.1.2",
-    "react": "^0.14.0 || >=15.0.0",
-    "react-dom": "^0.14.0 || >=15.0.0",
-    "style-loader": "~0.20.3",
+    "react": ">=16.3.0",
+    "react-dom": ">=16.3.0",
+    "react-styleguidist": "~8.0.6",
+    "resize-observer-polyfill": "~1.5.1",
+    "style-loader": "~0.23.1",
     "stylint": "~1.5.9",
-    "stylint-loader": "~1.0.0",
-    "stylus-loader": "~3.0.1",
-    "tap": "~11.1.2",
-    "trendmicro-ui": "~0.5.1",
-    "url-loader": "~0.6.2",
-    "webpack": "~3.10.0",
-    "webpack-dev-server": "~2.11.1"
+    "stylus-loader": "~3.0.2",
+    "tap": "~12.1.1",
+    "url-loader": "~1.1.2",
+    "webpack": "~4.28.1",
+    "webpack-cli": "~3.1.2",
+    "webpack-dev-server": "~3.1.10",
+    "which": "~1.3.1"
   }
 }
